requirements.txt
opencv-python>=4.8.0
PyQt6>=6.5.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pyzbar>=0.1.9
pyobjc-framework-AVFoundation>=10.0; sys_platform == 'darwin'
ultralytics>=8.0.0

